Fix booking flow, payment, avatar upload
- Fix EF Core column mismatches across all models (Booking, Driver, User, Payment, Notification, etc.)
- Fix SQL Server trigger conflict with UseSqlOutputClause(false)
- Add transaction to BookingService.CreateAsync to prevent orphan bookings
- Fix Stripe amount: convert VND to USD (÷25000) before creating PaymentIntent
- Add 'stripe' to Payment.payment_method DB constraint
- Fix Payment INSERT: set RegionId, PaymentStatusId, PaymentType
- Send confirmation email after booking created
- Add Cloudinary avatar upload (POST /api/users/me/avatar)
- Fix ProfilePage: read booking.carModel, wire up camera button
- Fix Stripe button display: show VND instead of USD

+++ car-rental/CarRental.API/DTOs/Common/RatingDto.cs
@@ -7,7 +7,7 @@ public class RatingDto
     public int CustomerId { get; set; }
     public string? CustomerName { get; set; }
     public int CarId { get; set; }
-    public decimal RatingScore { get; set; }
+    public byte RatingScore { get; set; }
     public string? Comment { get; set; }
     public DateTime CreatedAt { get; set; }
 }
@@ -16,6 +16,6 @@ public class CreateRatingRequest
 {
     public int BookingId { get; set; }
     public int CarId { get; set; }
-    public decimal RatingScore { get; set; }
+    public byte RatingScore { get; set; }
     public string? Comment { get; set; }
 }
